requirements.txt
streamlit==1.32.0
pandas==2.2.1
plotly==5.18.0
pillow==10.2.0
requests==2.31.0
python-dateutil==2.8.2